bench: isolate k-means performance cases
